feature/FZ5BR-59-KR add ResultsTable and ResultsItem to self-components
