+++ src/components/map-items/Results/adjustTime.js
@@ -8,7 +8,7 @@ export const adjustTime = (departures, type, inputTime) => {
             )
         )
     ) ?
-        departures.reduce((min, next) => Math.min(min, next), 0) :
+        departures.reduce((min, next) => min < next ? min : next, 0) :
         departures.filter(
             ({ depFromEndStop, depFromStartStop }) => (
                 (
@@ -17,5 +17,5 @@ export const adjustTime = (departures, type, inputTime) => {
                         depFromStartStop
                 ) >=  inputTime
             )
-        ).reduce((min, next) => Math.min(min, next), 0)
+        ).reduce((min, next) => min < next ? min : next, 0)
 }
